pyproject.toml
[tool.poetry]
name = "nonebot-plugin-auto-sendlike"
version = "0.0.9"
description = "Nonebot2 的点赞、订阅赞功能，每天 12 点定时点赞👍！轻量、高效、便捷的小插件！"
authors = ["zhiyu1998 <renzhiyu0416@qq.com>"]
readme = "README.md"
packages = [{include = "nonebot-plugin-auto-sendlike"}]
repository = "https://github.com/zhiyu1998/nonebot-plugin-auto-sendlike"
keywords = ["nonebot", "nonebot2", "resolver"]

[tool.poetry.dependencies]
python = ">=3.9"
nonebot2 = ">=2.3.0"
nonebot_plugin_apscheduler = "^0.5.0"
nonebot_plugin_localstore = "^0.7.1"
nonebot-adapter-onebot = ">=2.0.0-beta.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"